refactor(analytics): extract build-path composition into ChampionBuildPathBuilder (P10.1)
First cut of decomposing the ChampionAnalyticsComputeService god-file. The
self-contained build-path composition logic — the timing-aware build-path sections
(starters/boots/core+timing/situational), item-set normalization/grouping, and rune
info — moves verbatim into a dedicated ChampionBuildPathBuilder (takes the DbContext),
leaving the compute service to call into it. NormalizeCompletedBuildItems/BuildRuneInfo
are shared by ComputeBuilds/ComputeProBuilds, so they live on the builder and the
stay-code routes through it; the EF-projected record types are nested on the builder
(translate identically).

Pure, behavior-preserving move — no logic changed. ChampionAnalyticsComputeService.cs
1765 -> 1300 lines (-465); the build-path logic is a focused 487-line class.

Gate: solution builds clean; full Service.Core (183, incl. the build-path +
ChampionBuildSnapshot equivalence tests) + WebAPI (40) suites green. No schema change.
